requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.10.0
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
Django==3.2.9
django-background-tasks==1.2.5
django-compat==1.0.15
django-extensions==3.1.3
django-qr-code==2.2.0
frozenlist==1.2.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
lxml==4.6.3
multidict==5.2.0
priority==1.3.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
python-decouple==3.5
pytz==2021.3
segno==1.3.3
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
yarl==1.7.0
zope.interface==5.4.0
